+++ Juego/Piezas/Queen.py
@@ -2,4 +2,4 @@ from Piezas.Piece import Piece
 
 class Queen(Piece):
     def __init__(self, color):
-        super().__init__(color, "Queen", "Q")
+        super().__init__(color, "Queen", "♕" if color == "Black" else "♛")